+++ src/command/factory.ts
@@ -1,6 +1,8 @@
 import type { Context } from '@model/context';
 import type { AudioService } from '@service/audio';
 import type { AutoplayService } from '@service/autoplay';
+import type { BrailleService } from '@service/braille';
+import type { DisplayService } from '@service/display';
 import type { HighContrastService } from '@service/highContrast';
 import type { HighlightService } from '@service/highlight';
 import type { NotificationService } from '@service/notification';
@@ -52,6 +54,7 @@ import {
   GridCellMoveRightCommand,
 } from './gridCell';
 import {
+  ExitBrailleAndSubplotCommand,
   MoveDownCommand,
   MoveLeftCommand,
   MoveRightCommand,
@@ -98,6 +101,8 @@ export class CommandFactory {
 
   private readonly audioService: AudioService;
   private readonly autoplayService: AutoplayService;
+  private readonly brailleService: BrailleService;
+  private readonly displayService: DisplayService;
   private readonly highContrastService: HighContrastService;
   private readonly highlightService: HighlightService;
   private readonly notificationService: NotificationService;
@@ -123,6 +128,8 @@ export class CommandFactory {
 
     this.audioService = commandContext.audioService;
     this.autoplayService = commandContext.autoplayService;
+    this.brailleService = commandContext.brailleService;
+    this.displayService = commandContext.displayService;
     this.highContrastService = commandContext.highContrastService;
     this.highlightService = commandContext.highlightService;
     this.notificationService = commandContext.notificationService;
@@ -185,9 +192,11 @@ export class CommandFactory {
         return new MoveToRightExtremeCommand(this.context);
 
       case 'MOVE_TO_TRACE_CONTEXT':
-        return new MoveToTraceContextCommand(this.context);
+        return new MoveToTraceContextCommand(this.context, this.brailleService, this.displayService);
       case 'MOVE_TO_SUBPLOT_CONTEXT':
         return new MoveToSubplotContextCommand(this.context);
+      case 'EXIT_BRAILLE_AND_SUBPLOT':
+        return new ExitBrailleAndSubplotCommand(this.context, this.displayService);
       case 'MOVE_TO_NEXT_TRACE':
         return new MoveToNextTraceCommand(this.context);
       case 'MOVE_TO_PREV_TRACE':